+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Convert tensors to python numbers (gorarakelyan)
+
 ## 2.3.0 Apr 10 2021
 - Bump up Aim UI to v1.5.0 (gorarakelyan)
 - Set default interval of sys tracking to 10 seconds (gorarakelyan)